Google translation and TTS as functions
Did not re-format in original.
Note: There have been issues with encoding, when setting system locale to Chinese, or global environment encoding to UTF-8, it seems to wreck either the google translation or the newsapi text. But, the above example should run with R, no changes to locale or default R encoding.